backend/requirements.txt
fastapi>=0.128.0
uvicorn>=0.40.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.21
python-dotenv>=1.2.1
pydantic>=2.0.0
llama-index>=0.14.12
openai>=2.14.0
qdrant-client>=1.16.2
streamlit>=1.52.2
inngest>=0.5.13
